+++ doc/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 
 - fix #45 sometimes commands run forever
 - fix #44 minor bug when checking out from repo with default branch not master
+- fix #41 improvedAPI, get_vcs_client function part of vcstools module
 
 0.1.26
 ------